==> final_mini_project/youn/20191212/01_csv_create_smote.ipynb <==
import numpy as np
import pandas as pd
import tensorflow as tf
import sklearn
from sklearn.model_selection import train_test_split
from imblearn.over_sampling import SMOTE
print(tf.__version__)

tf.enable_eager_execution() 

df = pd.read_csv('./input/train.csv')
X_train = df.drop('Class', axis=1)
y_train = df.Class

print(df.Class.value_counts())

sm = SMOTE(random_state=27)
X_train_smote, y_train_smote = sm.fit_sample(X_train, y_train)
X_SMOTE = pd.concat([pd.DataFrame(X_train_smote), pd.DataFrame({'Class':y_train_smote})], axis=1)
X_SMOTE_shuffle = sklearn.utils.shuffle(X_SMOTE)
X_SMOTE.Class.value_counts()

y_train.value_counts()

y_train_smote.value_counts()

X_SMOTE_shuffle.Class.value_counts()

X_SMOTE_shuffle.to_csv('./input/train_smote.csv',index=False)

X_SMOTE_shuffle.head()

X_SMOTE_shuffle.tail()
