==> final_mini_project/youn/20191212/01_csv_create_time_minus.ipynb <==
import numpy as np
import pandas as pd
import tensorflow as tf
import sklearn
from sklearn.model_selection import train_test_split
from imblearn.over_sampling import SMOTE
print(tf.__version__)

tf.enable_eager_execution() 

df = pd.read_csv('./input/train.csv')
df = df.drop('Time', axis=1)
df = df.drop('Amount', axis=1)

print(df.Class.value_counts())

df

df.to_csv('/home/jupyter/google-asl-study/final_mini_project/youn/input/train_time_amount_minus.csv',index=False)

df = pd.read_csv('./input/train_smote.csv')
df = df.drop('Time', axis=1)
df = df.drop('Amount', axis=1)

df.to_csv('/home/jupyter/google-asl-study/final_mini_project/youn/input/train_smote_time_amount_minus.csv',index=False)

df = pd.read_csv('./input/valid.csv')
df = df.drop('Time', axis=1)
df = df.drop('Amount', axis=1)

df.to_csv('/home/jupyter/google-asl-study/final_mini_project/youn/input/valid_time_amount_minus.csv',index=False)

df = pd.read_csv('./input/test.csv')
df = df.drop('Time', axis=1)
df = df.drop('Amount', axis=1)

df.to_csv('/home/jupyter/google-asl-study/final_mini_project/youn/input/test_time_amount_minus.csv',index=False)
