==> final_mini_project/youn/20191212/01_csv_small_create.ipynb <==
import numpy as np
import tensorflow as tf

tf.enable_eager_execution() 

CSV_COLUMN_NAMES = ["time"]+['V'+(i+1).astype(np.str) for i in np.array(range(28))]+['Amount','Class']
CSV_DEFAULTS = [[0.0] for i in np.array(range(30))]+[[0]]
print('CSV_COLUMN_NAMES:{}\nCSV_DEFAULTS:{}'.format(CSV_COLUMN_NAMES,CSV_DEFAULTS))

def parse_row(row):
    fields = tf.decode_csv(row, record_defaults=CSV_DEFAULTS)
    features = {    
                'time':fields[0], 
                'V1'  :fields[1],                 'V2'  :fields[2],                 'V3'  :fields[3], 
                'V4'  :fields[4],                 'V5'  :fields[5],                 'V6'  :fields[6], 
                'V7'  :fields[7],                 'V8'  :fields[8],                 'V9'  :fields[9], 
                'V10' :fields[10],                'V11' :fields[11],                'V12' :fields[12],
                'V13' :fields[13],                'V14' :fields[14],                'V15' :fields[15], 
                'V16' :fields[16],                'V17' :fields[17],                'V18' :fields[18], 
                'V19' :fields[19],                'V20' :fields[20],                'V21' :fields[21], 
                'V22' :fields[22],                'V23' :fields[23],                'V24' :fields[24], 
                'V25' :fields[25],                'V26' :fields[26],                'V27' :fields[27], 
                'V28' :fields[28],                 'Amount':fields[29]}
    labels = fields[30]
    
    return features, labels

def read_dataset(csv_path):  
    dataset = tf.data.TextLineDataset(csv_path)
    dataset = dataset.skip(1).map(parse_row).shuffle(1000000)
    return dataset

!head ./input/creditcard.csv > ./input/creditcard_small.csv

for i, (feature, label) in enumerate(read_dataset("./input/creditcard_eval.csv")):
    if i >=2000:
        break
    if label.numpy() == 1:
        print("time:{}, V1:{}, V2:{}, V28:{}, Amount:{}, Class:{}".format(feature["time"].numpy(),
                                                                      feature["V1"].numpy(),
                                                                      feature["V2"].numpy(),
                                                                      feature["V28"].numpy(),
                                                                      feature["Amount"].numpy(),
                                                                      label.numpy()))
#    print("*"*100)
#    print("time  :", feature["time"].numpy())
#    print("V1    :", feature["V1"].numpy())
#    print("V2    :", feature["V2"].numpy())
#    print("V3    :", feature["V3"].numpy())
#    print("V27   :", feature["V27"].numpy())
#    print("V28   :", feature["V28"].numpy())
#    print("Amount:", feature["Amount"].numpy())
#    print("Class :", label.numpy())
print("*"*100)

def train_input_fn(csv_path, batch_size = 128):
    dataset = tf.data.TextLineDataset(csv_path)   \
                             .map(parse_row)
    dataset = dataset.shuffle(1000)    \
                     .repeat(15)    \
                     .batch(128)
    return dataset
