==> final_mini_project/youn/20191212/01_csv_train_eval_create.ipynb <==
import numpy as np
import pandas as pd
import tensorflow as tf
from sklearn.model_selection import train_test_split

df = pd.read_csv('./input/creditcard.csv')
df_train, df_test = train_test_split(df, test_size=0.25, random_state=27)
print(df_train.Class.value_counts())
print(df_test.Class.value_counts())

!head ./input/creditcard.csv > ./input/creditcard_small.csv
!head -5 ./input/creditcard_small.csv

df_train.to_csv('./input/creditcard_train.csv',index=False)

df_test.to_csv('./input/creditcard_eval.csv',index=False)

!head -5 ./input/creditcard_train.csv

!head -5 ./input/creditcard_eval.csv
