==> final_mini_project/youn/20191212/03_estimator_자체제작_dnnregressor_20191208.ipynb <==
import tensorflow as tf
import numpy as np
import shutil
print(tf.__version__)

CSV_COLUMN_NAMES = ["time"]+['V'+(i+1).astype(np.str) for i in np.array(range(28))]+['Amount','Class']
CSV_DEFAULTS = [[0.0] for i in np.array(range(30))]+[[0]]
print('CSV_COLUMN_NAMES:{}\nCSV_DEFAULTS:{}'.format(CSV_COLUMN_NAMES,CSV_DEFAULTS))

FEATURE_NAMES = CSV_COLUMN_NAMES[:-1] # all but first column

feature_cols = [tf.feature_column.numeric_column(key = k) for k in FEATURE_NAMES]
for i in feature_cols:
    print(i)  

def serving_input_receiver_fn():
    receiver_tensors = {'time'  :tf.placeholder(dtype = tf.float32,shape=[None]),
                        'V1'    :tf.placeholder(dtype = tf.float32,shape=[None]),
                        'V2'    :tf.placeholder(dtype = tf.float32,shape=[None]),
                        'V3'    :tf.placeholder(dtype = tf.float32,shape=[None]),
                        'V4'    :tf.placeholder(dtype = tf.float32,shape=[None]),
                        'V5'    :tf.placeholder(dtype = tf.float32,shape=[None]),
                        'V6'    :tf.placeholder(dtype = tf.float32,shape=[None]),
                        'V7'    :tf.placeholder(dtype = tf.float32,shape=[None]),
                        'V8'    :tf.placeholder(dtype = tf.float32,shape=[None]),
                        'V9'    :tf.placeholder(dtype = tf.float32,shape=[None]),
                        'V10'   :tf.placeholder(dtype = tf.float32,shape=[None]),
                        'V11'   :tf.placeholder(dtype = tf.float32,shape=[None]),
                        'V12'   :tf.placeholder(dtype = tf.float32,shape=[None]),
                        'V13'   :tf.placeholder(dtype = tf.float32,shape=[None]),
                        'V14'   :tf.placeholder(dtype = tf.float32,shape=[None]),
                        'V15'   :tf.placeholder(dtype = tf.float32,shape=[None]),
                        'V16'   :tf.placeholder(dtype = tf.float32,shape=[None]),
                        'V17'   :tf.placeholder(dtype = tf.float32,shape=[None]),
                        'V18'   :tf.placeholder(dtype = tf.float32,shape=[None]),
                        'V19'   :tf.placeholder(dtype = tf.float32,shape=[None]),
                        'V20'   :tf.placeholder(dtype = tf.float32,shape=[None]),
                        'V21'   :tf.placeholder(dtype = tf.float32,shape=[None]),
                        'V22'   :tf.placeholder(dtype = tf.float32,shape=[None]),
                        'V23'   :tf.placeholder(dtype = tf.float32,shape=[None]),
                        'V24'   :tf.placeholder(dtype = tf.float32,shape=[None]),
                        'V25'   :tf.placeholder(dtype = tf.float32,shape=[None]),
                        'V26'   :tf.placeholder(dtype = tf.float32,shape=[None]),
                        'V27'   :tf.placeholder(dtype = tf.float32,shape=[None]),
                        'V28'   :tf.placeholder(dtype = tf.float32,shape=[None]),
                        'Amount':tf.placeholder(dtype = tf.float32,shape=[None])
                        }
    features = receiver_tensors
    return tf.estimator.export.ServingInputReceiver(features = features, receiver_tensors = receiver_tensors)


def parse_row(row):
    fields = tf.decode_csv(records = row, record_defaults = CSV_DEFAULTS)
    features = dict(zip(CSV_COLUMN_NAMES, fields))
    label = features.pop("Class")
    return features, label

def read_dataset(csv_path):
    dataset = tf.data.TextLineDataset(filenames = csv_path).skip(count = 1) # skip header
    dataset = dataset.map(map_func = parse_row)
    return dataset

def train_input_fn(csv_path, batch_size = 128):
    dataset = read_dataset(csv_path)
    dataset = dataset.shuffle(buffer_size = 1000).repeat(count = None).batch(batch_size = batch_size)
    return dataset

def eval_input_fn(csv_path, batch_size = 128):
    dataset = read_dataset(csv_path)
    dataset = dataset.batch(batch_size = batch_size)
    return dataset

def my_auc(labels, predictions, features):
    auc_metric = tf.keras.metrics.AUC(name="my_auc")

    #dict_keys(['probabilities', 'class_ids', 'logits', 'all_classes', 'classes', 'logistic', 'all_class_ids'])    

    auc_metric.update_state(y_true=labels, y_pred=predictions['logistic'])
    return {'auc': auc_metric}

def custom_metric(labels, predictions):
    print('>>>>',predictions.keys())
    predicted_classes = predictions["predictions"]  
    custom_metric = tf.metrics.mean(labels, predicted_classes, name="custom_metric")
    return {"custom_metric": custom_metric}


OUTDIR = "credit_trained"

config = tf.estimator.RunConfig(
    model_dir=OUTDIR,
    save_summary_steps=100,
    save_checkpoints_steps=2000
)

#myopt = tf.train.AdamOptimizer(learning_rate=0.01)
model = tf.estimator.DNNClassifier(
                                  #model_dir=OUTDIR,
                                  hidden_units=[10, 10],
                                  feature_columns=feature_cols,
                                  #optimizer=myopt,
                                  config = config)
#model = tf.estimator.add_metrics(model, my_auc)

!rm -r credit_trained

model.train(input_fn = lambda : train_input_fn('./input/creditcard_train.csv'),
    max_steps = 500
)

train_spec = tf.estimator.TrainSpec(
    input_fn = lambda: train_input_fn('./input/creditcard_train.csv'),
    max_steps = 500
)
exporter = tf.estimator.FinalExporter('./exporter', serving_input_receiver_fn=serving_input_receiver_fn)
eval_spec = tf.estimator.EvalSpec(
    input_fn = lambda: eval_input_fn("./input/creditcard_eval.csv"),
    steps = None,
    start_delay_secs = 1,
    throttle_secs = 1,
    exporters = exporter,
)

!rm -r credit_trained

tf.estimator.train_and_evaluate(estimator = model, 
                                train_spec = train_spec, 
                                eval_spec = eval_spec)
