==> final_mini_project/youn/20191212/05_dnnclassifier_smote_rate_0001_20191211_1400_lhb.ipynb <==
import tensorflow as tf
import numpy as np
import pandas as pd
import shutil
import os
import sys
import seaborn as sns
import matplotlib.pyplot as plt

print(tf.__version__)

tf.enable_eager_execution() 

CSV_COLUMN_NAMES = ["Time"]+['V'+(i+1).astype(np.str) for i in np.array(range(28))]+['Amount','Class']
CSV_DEFAULTS = [[0.0] for i in np.array(range(30))]+[[0]]
print('CSV_COLUMN_NAMES:{}\nCSV_DEFAULTS:{}'.format(CSV_COLUMN_NAMES,CSV_DEFAULTS))

FEATURE_NAMES = CSV_COLUMN_NAMES[1:-1] # all but first column
feature_cols = [tf.feature_column.numeric_column(key = k) for k in FEATURE_NAMES]
for i in feature_cols:
    print(i)  

def parse_row(row):
    fields = tf.decode_csv(records = row, record_defaults = CSV_DEFAULTS)
    features = dict(zip(CSV_COLUMN_NAMES, fields))
    features.pop("Time")
    label = features.pop("Class")
    return features, label

def read_dataset(csv_path):
    dataset = tf.data.TextLineDataset(filenames = csv_path).skip(count = 1) # skip header
    dataset = dataset.map(map_func = parse_row)
    return dataset

def train_input_fn(csv_path, batch_size = 128):
    dataset = read_dataset(csv_path)
    dataset = dataset.shuffle(buffer_size = 1000).repeat(count = None).batch(batch_size = batch_size)
    return dataset

def eval_input_fn(csv_path, batch_size = 128):
    dataset = read_dataset(csv_path)
    dataset = dataset.batch(batch_size = batch_size)
    return dataset

# 실제 데이터 확인
# tf.enable_eager_execution() 선언되어 있어야 함.
#train_data_check = train_input_fn(csv_path = './input/train_smote.csv')
#print(train_data_check.make_one_shot_iterator().get_next()[0]["Amount"][:10])
#print(train_data_check.make_one_shot_iterator().get_next()[1][:10])


def get_predictions(estimator, input_fn):
    return [x["class_ids"][0] for x in model.predict(input_fn=input_fn)]

def confusion_display(model, filename):
    LABELS = ["0", "1"]

    df_eval = pd.read_csv(filename)

    # Create a confusion matrix on training data.
    with tf.Graph().as_default():
        cm = tf.confusion_matrix(df_eval["Class"],get_predictions(model,  lambda: eval_input_fn(filename)))
        with tf.Session() as session:
            cm_out = session.run(cm)
            print(cm_out)

    # Normalize the confusion matrix so that each row sums to 1.
    cm_out = cm_out.astype(float) 
    # / cm_out.sum(axis=1)[:, np.newaxis]

    print()
    print('Class1')
    print('precision :',cm_out[1][1]/(cm_out[0][1]+cm_out[1][1]))
    print('recall :',cm_out[1][1]/(cm_out[1][0]+cm_out[1][1]))

    #sns.heatmap(cm_out, annot=True, xticklabels=LABELS, yticklabels=LABELS);
    #plt.xlabel("Predicted");
    #plt.ylabel("True");

def outdir_fn(OUTDIR, unit):
    for i, j in enumerate(unit):
        OUTDIR = OUTDIR + "_" + str(j)
    return OUTDIR

unit = [128,64,32,32]
OUTDIR = "credit_trained_LR0001"
OUTDIR = outdir_fn(OUTDIR, unit)

config = tf.estimator.RunConfig(
    model_dir=OUTDIR,
#    tf_random_seed = 1,
    keep_checkpoint_max=5,
    save_summary_steps=100,
    save_checkpoints_steps=1000
)
model = tf.estimator.DNNClassifier(
                                  hidden_units=unit,
                                  feature_columns=feature_cols,
                                  optimizer = tf.train.AdamOptimizer(learning_rate=0.001),
                                  dropout=0.2,
                                  config = config)

os.system ('rm -r {}'.format(OUTDIR))
model.train(input_fn = lambda : train_input_fn('./input/train_lhb.csv'), max_steps = 2000)
model.evaluate(input_fn = lambda: eval_input_fn('./input/valid.csv'))
confusion_display(model, './input/valid.csv')

unit = [128,64,32,32]
OUTDIR = "credit_trained_LR0001"
OUTDIR = outdir_fn(OUTDIR, unit)

config = tf.estimator.RunConfig(
    model_dir=OUTDIR,
#    tf_random_seed = 1,
    keep_checkpoint_max=5,
    save_summary_steps=100,
    save_checkpoints_steps=1000
)
model = tf.estimator.DNNClassifier(
                                  hidden_units=unit,
                                  feature_columns=feature_cols,
                                  optimizer = tf.train.AdamOptimizer(learning_rate=0.001),
                                  dropout=0.2,
                                  config = config)

os.system ('rm -r {}'.format(OUTDIR))
model.train(input_fn = lambda : train_input_fn('./input/train.csv'), max_steps = 2000)
model.evaluate(input_fn = lambda: eval_input_fn('./input/valid.csv'))
confusion_display(model, './input/valid.csv')

unit = [30]
OUTDIR = "credit_trained_LR0001"
OUTDIR = outdir_fn(OUTDIR, unit)

config = tf.estimator.RunConfig(
    model_dir=OUTDIR,
#    tf_random_seed = 1,
    keep_checkpoint_max=5,
    save_summary_steps=100,
    save_checkpoints_steps=1000
)
model = tf.estimator.DNNClassifier(
                                  hidden_units=unit,
                                  feature_columns=feature_cols,
                                  optimizer = tf.train.AdamOptimizer(learning_rate=0.001),
                                  dropout=0.2,
                                  config = config)

os.system ('rm -r {}'.format(OUTDIR))
model.train(input_fn = lambda : train_input_fn('./input/train_smote.csv'), max_steps = 2000)
model.evaluate(input_fn = lambda: eval_input_fn('./input/valid.csv'))
confusion_display(model, './input/valid.csv')

unit = [30,10]
OUTDIR = "credit_trained_LR0001"
OUTDIR = outdir_fn(OUTDIR, unit)

config = tf.estimator.RunConfig(
    model_dir=OUTDIR,
#    tf_random_seed = 1,
    keep_checkpoint_max=5,
    save_summary_steps=100,
    save_checkpoints_steps=1000
)
model = tf.estimator.DNNClassifier(
                                  hidden_units=unit,
                                  feature_columns=feature_cols,
                                  optimizer = tf.train.AdamOptimizer(learning_rate=0.001),
                                  dropout=0.2,
                                  config = config)

os.system ('rm -r {}'.format(OUTDIR))
model.train(input_fn = lambda : train_input_fn('./input/train_smote.csv'), max_steps = 2000)
model.evaluate(input_fn = lambda: eval_input_fn('./input/valid.csv'))
confusion_display(model, './input/valid.csv')

unit = [30,10,30]
OUTDIR = "credit_trained_LR0001"
OUTDIR = outdir_fn(OUTDIR, unit)

config = tf.estimator.RunConfig(
    model_dir=OUTDIR,
#    tf_random_seed = 1,
    keep_checkpoint_max=5,
    save_summary_steps=100,
    save_checkpoints_steps=1000
)
model = tf.estimator.DNNClassifier(
                                  hidden_units=unit,
                                  feature_columns=feature_cols,
                                  optimizer = tf.train.AdamOptimizer(learning_rate=0.001),
                                  dropout=0.2,
                                  config = config)

os.system ('rm -r {}'.format(OUTDIR))
model.train(input_fn = lambda : train_input_fn('./input/train_smote.csv'), max_steps = 2000)
model.evaluate(input_fn = lambda: eval_input_fn('./input/valid.csv'))
confusion_display(model, './input/valid.csv')

unit = [30,100]
OUTDIR = "credit_trained_LR0001"
OUTDIR = outdir_fn(OUTDIR, unit)

config = tf.estimator.RunConfig(
    model_dir=OUTDIR,
#    tf_random_seed = 1,
    keep_checkpoint_max=5,
    save_summary_steps=100,
    save_checkpoints_steps=1000
)
model = tf.estimator.DNNClassifier(
                                  hidden_units=unit,
                                  feature_columns=feature_cols,
                                  optimizer = tf.train.AdamOptimizer(learning_rate=0.001),
                                  dropout=0.2,
                                  config = config)

os.system ('rm -r {}'.format(OUTDIR))
model.train(input_fn = lambda : train_input_fn('./input/train_smote.csv'), max_steps = 2000)
model.evaluate(input_fn = lambda: eval_input_fn('./input/valid.csv'))
confusion_display(model, './input/valid.csv')

unit = [100,10]
OUTDIR = "credit_trained_LR0001"
OUTDIR = outdir_fn(OUTDIR, unit)

config = tf.estimator.RunConfig(
    model_dir=OUTDIR,
    tf_random_seed = 1,
    keep_checkpoint_max=5,
    save_summary_steps=100,
    save_checkpoints_steps=1000
)
model = tf.estimator.DNNClassifier(
                                  hidden_units=unit,
                                  feature_columns=feature_cols,
                                  optimizer = tf.train.AdamOptimizer(learning_rate=0.001),
                                  dropout=0.2,
                                  config = config)

os.system ('rm -r {}'.format(OUTDIR))
model.train(input_fn = lambda : train_input_fn('./input/train_smote.csv'), max_steps = 2000)
model.evaluate(input_fn = lambda: eval_input_fn('./input/valid.csv'))
confusion_display(model, './input/valid.csv')

unit = [128,64,32,10]
OUTDIR = "credit_trained_LR0001"
OUTDIR = outdir_fn(OUTDIR, unit)

config = tf.estimator.RunConfig(
    model_dir=OUTDIR,
#    tf_random_seed = 1,
    keep_checkpoint_max=5,
    save_summary_steps=100,
    save_checkpoints_steps=1000
)
model = tf.estimator.DNNClassifier(
                                  hidden_units=unit,
                                  feature_columns=feature_cols,
                                  optimizer = tf.train.AdamOptimizer(learning_rate=0.001),
                                  dropout=0.2,
                                  config = config)

os.system ('rm -r {}'.format(OUTDIR))
model.train(input_fn = lambda : train_input_fn('./input/train.csv'), max_steps = 2000)
model.evaluate(input_fn = lambda: eval_input_fn('./input/valid.csv'))
confusion_display(model, './input/valid.csv')

unit = [128,64,32,5]
OUTDIR = "credit_trained_LR0001"
OUTDIR = outdir_fn(OUTDIR, unit)

config = tf.estimator.RunConfig(
    model_dir=OUTDIR,
#    tf_random_seed = 1,
    keep_checkpoint_max=5,
    save_summary_steps=100,
    save_checkpoints_steps=1000
)
model = tf.estimator.DNNClassifier(
                                  hidden_units=unit,
                                  feature_columns=feature_cols,
                                  optimizer = tf.train.AdamOptimizer(learning_rate=0.001),
                                  dropout=0.2,
                                  config = config)

os.system ('rm -r {}'.format(OUTDIR))
model.train(input_fn = lambda : train_input_fn('./input/train.csv'), max_steps = 2000)
model.evaluate(input_fn = lambda: eval_input_fn('./input/valid.csv'))
confusion_display(model, './input/valid.csv')

unit = [128,32,5]
OUTDIR = "credit_trained_LR0001"
OUTDIR = outdir_fn(OUTDIR, unit)

config = tf.estimator.RunConfig(
    model_dir=OUTDIR,
#    tf_random_seed = 1,
    keep_checkpoint_max=5,
    save_summary_steps=100,
    save_checkpoints_steps=1000
)
model = tf.estimator.DNNClassifier(
                                  hidden_units=unit,
                                  feature_columns=feature_cols,
                                  optimizer = tf.train.AdamOptimizer(learning_rate=0.001),
                                  dropout=0.2,
                                  config = config)

os.system ('rm -r {}'.format(OUTDIR))
model.train(input_fn = lambda : train_input_fn('./input/train.csv'), max_steps = 2000)
model.evaluate(input_fn = lambda: eval_input_fn('./input/valid.csv'))
confusion_display(model, './input/valid.csv')

unit = [128,64,5]
OUTDIR = "credit_trained_LR0001"
OUTDIR = outdir_fn(OUTDIR, unit)

config = tf.estimator.RunConfig(
    model_dir=OUTDIR,
#    tf_random_seed = 1,
    keep_checkpoint_max=5,
    save_summary_steps=100,
    save_checkpoints_steps=1000
)
model = tf.estimator.DNNClassifier(
                                  hidden_units=unit,
                                  feature_columns=feature_cols,
                                  optimizer = tf.train.AdamOptimizer(learning_rate=0.001),
                                  dropout=0.2,
                                  config = config)

os.system ('rm -r {}'.format(OUTDIR))
model.train(input_fn = lambda : train_input_fn('./input/train.csv'), max_steps = 2000)
model.evaluate(input_fn = lambda: eval_input_fn('./input/valid.csv'))
confusion_display(model, './input/valid.csv')

unit = [128,32,10]
OUTDIR = "credit_trained_LR0001"
OUTDIR = outdir_fn(OUTDIR, unit)

config = tf.estimator.RunConfig(
    model_dir=OUTDIR,
#    tf_random_seed = 1,
    keep_checkpoint_max=5,
    save_summary_steps=100,
    save_checkpoints_steps=1000
)
model = tf.estimator.DNNClassifier(
                                  hidden_units=unit,
                                  feature_columns=feature_cols,
                                  optimizer = tf.train.AdamOptimizer(learning_rate=0.001),
                                  dropout=0.2,
                                  config = config)

os.system ('rm -r {}'.format(OUTDIR))
model.train(input_fn = lambda : train_input_fn('./input/train.csv'), max_steps = 2000)
model.evaluate(input_fn = lambda: eval_input_fn('./input/valid.csv'))
confusion_display(model, './input/valid.csv')

unit = [128,64,10]
OUTDIR = "credit_trained_LR0001"
OUTDIR = outdir_fn(OUTDIR, unit)

config = tf.estimator.RunConfig(
    model_dir=OUTDIR,
#    tf_random_seed = 1,
    keep_checkpoint_max=5,
    save_summary_steps=100,
    save_checkpoints_steps=1000
)
model = tf.estimator.DNNClassifier(
                                  hidden_units=unit,
                                  feature_columns=feature_cols,
                                  optimizer = tf.train.AdamOptimizer(learning_rate=0.001),
                                  dropout=0.2,
                                  config = config)

os.system ('rm -r {}'.format(OUTDIR))
model.train(input_fn = lambda : train_input_fn('./input/train.csv'), max_steps = 2000)
model.evaluate(input_fn = lambda: eval_input_fn('./input/valid.csv'))
confusion_display(model, './input/valid.csv')
