==> final_mini_project/youn/20191212/99_Tensorboard.ipynb <==
OUTDIR='/home/jupyter/google-asl-study/final_mini_project/youn/trained_dnnclass/miniproj'
#OUTDIR='gs://qwiklabs-gcp-ml-49b827b781ab/miniproj/dnnclass_HParam/4'

get_ipython().system_raw(
    "tensorboard --logdir {} --host 0.0.0.0 --port 6006 &"
    .format(OUTDIR)
)

get_ipython().system_raw("/home/jupyter/training-data-analyst/courses/machine_learning/deepdive/02_tensorflow/assets/ngrok http 6006 &")

!curl -s http://localhost:4040/api/tunnels | python3 -c \
    "import sys, json; print(json.load(sys.stdin)['tunnels'][0]['public_url'])"

!curl -s http://localhost:4040/api/tunnels | python3 -c     "import sys, json; print(json.load(sys.stdin)['tunnels'][0]['public_url'])"

### Tensorboard cleanup

# this will kill the processes for Tensorboard
!ps aux | grep tensorboard | awk '{print $2}' | xargs kill

# this will kill the processes for ngrok
!ps aux | grep ngrok | awk '{print $2}' | xargs kill
